+++ app/Http/Controllers/ReservationController.php
@@ -28,7 +28,15 @@ class ReservationController extends Controller
 
     public function all()
     {
-        $reservations = Reservation::with(['vehicule', 'client', 'user'])->latest()->paginate(10);
+        $user = auth()->user();
+        if ($user->isAdmin()) {
+            $reservations = Reservation::with(['vehicule', 'client', 'user'])->latest()->paginate(10);
+        }  else {
+            $reservations = $user->reservations()
+            ->with(['vehicule', 'client', 'user'])
+            ->latest()
+            ->paginate(1);
+        }
         return response()->json(['code' => 200, 'reservations' => $reservations], 200);
     }
 
@@ -72,23 +80,26 @@ class ReservationController extends Controller
      */
     public function update(ReservationRequest $request, $id)
     {
-        // Validation
-        $formData = $request->validated();
-        $formData['user_id'] = auth()->user()->id;
-        $dateDebut = [
-            'date_debut' => DateTimeHelper::separateDateTime($request->date_debut)['date'],
-            'heure_debut' => DateTimeHelper::separateDateTime($request->date_debut)['time'],
-        ];
-        $dateFin = [
-            'date_fin' => DateTimeHelper::separateDateTime($request->date_fin)['date'],
-            'heure_fin' => DateTimeHelper::separateDateTime($request->date_fin)['time'],
-        ];
-        $formData = array_merge($formData, $dateDebut, $dateFin);
         $reservation = Reservation::find($id);
         if ($reservation) {
-            $reservation->update($formData);
-            // Return success response if data is updated successfully
-            return response()->json(['status' => 200, 'msg' => 'Opération effectuée avec succès.'], 200); 
+            if ($reservation->user_id === auth()->user()->id || auth()->user()->isAdmin()) {
+                // Validation
+                $formData = $request->validated();
+                $dateDebut = [
+                    'date_debut' => DateTimeHelper::separateDateTime($request->date_debut)['date'],
+                    'heure_debut' => DateTimeHelper::separateDateTime($request->date_debut)['time'],
+                ];
+                $dateFin = [
+                    'date_fin' => DateTimeHelper::separateDateTime($request->date_fin)['date'],
+                    'heure_fin' => DateTimeHelper::separateDateTime($request->date_fin)['time'],
+                ];
+                $formData = array_merge($formData, $dateDebut, $dateFin);
+                $reservation->update($formData);
+                // Return success response if data is updated successfully
+                return response()->json(['status' => 200, 'msg' => 'Opération effectuée avec succès.'], 200);
+            } else {
+                return response()->json(['status' => 401,'msg' => "vous n'êtes pas autorisé(e) à effectuer cette action."], 401);
+            }
         } 
         return response()->json(['status' => 404,'msg' => "Cette information n'existe pas"], 404);
     }
@@ -109,8 +120,12 @@ class ReservationController extends Controller
     {
         $reservation = Reservation::find($id);
         if ($reservation) {
-            $reservation->delete();
-            return response()->json(['status' => 200, 'success' => 'Opération effectuée avec succès.'], 200);
+            if ($reservation->user_id === auth()->user()->id || auth()->user()->isAdmin()) { 
+                $reservation->delete();
+                return response()->json(['status' => 200, 'success' => 'Opération effectuée avec succès.'], 200);
+            } else {
+                return response()->json(['status' => 401,'msg' => "vous n'êtes pas autorisé(e) à effectuer cette action."], 401);
+            }
         }
         return response()->json(['status' => 404,'msg' => "Cette information n'existe pas"], 404);
     }
@@ -121,21 +136,43 @@ class ReservationController extends Controller
     public function search(Request $request)
     {
       $value = $request->search;
-      $result = Reservation::with(['vehicule', 'client', 'user'])
-                                  ->whereHas('vehicule', function ($query) use ($value) {
+      // Search in all records
+      if (auth()->user()->isAdmin()) { 
+        $result = Reservation::with(['vehicule', 'client', 'user'])
+                              ->whereHas('vehicule', function ($query) use ($value) {
+                                    $query->where('matricule', 'LIKE', "%$value%")
+                                          ->orWhere('marque', 'LIKE', "%$value%");
+                                })
+                              ->orWhereHas('client', function ($query) use ($value) {
+                                $query->where('nom', 'LIKE', "%$value%")
+                                      ->orWhere('prenom', 'LIKE', "%$value%");
+                              })
+                              ->orWhereHas('user', function ($query) use ($value) {
+                                $query->where('nom', 'LIKE', "%$value%")
+                                      ->orWhere('prenom', 'LIKE', "%$value%");
+                              })
+                              ->latest()
+                              ->paginate(10);
+      } else { 
+        // Search only in records made by the logged in user
+        $result = auth()->user()->reservations()->with(['vehicule', 'client', 'user'])
+                                ->where(function ($query) use ($value) {
+                                    $query->whereHas('vehicule', function ($query) use ($value) {
                                         $query->where('matricule', 'LIKE', "%$value%")
-                                              ->orWhere('marque', 'LIKE', "%$value%");
+                                            ->orWhere('marque', 'LIKE', "%$value%");
                                     })
-                                  ->orWhereHas('client', function ($query) use ($value) {
-                                    $query->where('nom', 'LIKE', "%$value%")
-                                          ->orWhere('prenom', 'LIKE', "%$value%");
-                                  })
-                                  ->orWhereHas('user', function ($query) use ($value) {
-                                    $query->where('nom', 'LIKE', "%$value%")
-                                          ->orWhere('prenom', 'LIKE', "%$value%");
-                                  })
-                                  ->latest()
-                                  ->paginate(1);
+                                    ->orWhereHas('client', function ($query) use ($value) {
+                                        $query->where('nom', 'LIKE', "%$value%")
+                                            ->orWhere('prenom', 'LIKE', "%$value%");
+                                    })
+                                    ->orWhereHas('user', function ($query) use ($value) {
+                                        $query->where('nom', 'LIKE', "%$value%")
+                                            ->orWhere('prenom', 'LIKE', "%$value%");
+                                    });
+                                })
+                                ->latest()
+                                ->paginate(10);
+      }
       $result->appends($request->all());
       return response()->json(['code' => 200, "reservations" => $result], 200);
     }
@@ -146,6 +183,9 @@ class ReservationController extends Controller
       $dateFin = $request->date_fin;
       $status = $request->status;
       $reservations = Reservation::with(['vehicule', 'client', 'user'])
+      ->when(!auth()->user()->isAdmin(), function ($query) {
+        $query->where('user_id', auth()->user()->id);
+      })
       ->when($dateDebut && $dateFin && $status, function ($query) use ($dateDebut, $dateFin, $status) {
           $query->where('date_debut', '>=', $dateDebut)
                 ->where('date_fin', '<=', $dateFin)
